Changed types for id's to Integer. Changed the logic of session creation to make it more flexible for testing. Implemented some test cases.

+++ src/main/java/org/anch/converter/YearAttributeConverter.java
@@ -6,18 +6,18 @@ import jakarta.persistence.Converter;
 import java.time.Year;
 
 @Converter(autoApply = true)
-public class YearAttributeConverter implements AttributeConverter<Year, Short> {
+public class YearAttributeConverter implements AttributeConverter<Year, Integer> {
 
     @Override
-    public Short convertToDatabaseColumn(Year attribute) {
+    public Integer convertToDatabaseColumn(Year attribute) {
         if (attribute != null) {
-            return (short) attribute.getValue();
+            return attribute.getValue();
         }
         return null;
     }
 
     @Override
-    public Year convertToEntityAttribute(Short dbData) {
+    public Year convertToEntityAttribute(Integer dbData) {
         if (dbData != null) {
             return Year.of(dbData);
         }
