+++ src/main/java/org/anch/dao/InventoryDao.java
@@ -5,5 +5,5 @@ import org.anch.entity.Inventory;
 import java.util.List;
 
 public interface InventoryDao extends CrudDao<Integer, Inventory> {
-    List<Inventory> findAllInventoriesByFilmId(Short filmId, Byte storeId);
+    List<Inventory> findAllInventoriesByFilmId(Integer filmId, Integer storeId);
 }
