+++ src/main/java/org/anch/dao/LanguageDao.java
@@ -2,5 +2,5 @@ package org.anch.dao;
 
 import org.anch.entity.Language;
 
-public interface LanguageDao extends CrudDao<Byte, Language> {
+public interface LanguageDao extends CrudDao<Integer, Language> {
 }
