+++ src/main/java/org/anch/dao/StaffDao.java
@@ -2,5 +2,5 @@ package org.anch.dao;
 
 import org.anch.entity.Staff;
 
-public interface StaffDao extends CrudDao<Byte, Staff> {
+public interface StaffDao extends CrudDao<Integer, Staff> {
 }
